fix names of types

+++ ft_memcmp.c
@@ -6,15 +6,17 @@
 /*   By: gpires-c <gpires-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/26 19:34:47 by gpires-c          #+#    #+#             */
-/*   Updated: 2026/05/27 21:10:27 by gpires-c         ###   ########.fr       */
+/*   Updated: 2026/05/29 19:44:45 by gpires-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*p1;
 	unsigned char	*p2;
-	unsigned int	i;
+	size_t			i;
 
 	p1 = (unsigned char *)s1;
 	p2 = (unsigned char *)s2;
