+++ ft_memset.c
@@ -6,14 +6,16 @@
 /*   By: gpires-c <gpires-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/24 21:01:16 by gpires-c          #+#    #+#             */
-/*   Updated: 2026/05/27 20:24:14 by gpires-c         ###   ########.fr       */
+/*   Updated: 2026/05/29 19:53:47 by gpires-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *buffer, int c, unsigned int size)
+#include "libft.h"
+
+void	*ft_memset(void *buffer, int c, size_t size)
 {
 	unsigned char	*p;
-	unsigned int	i;
+	size_t			i;
 
 	i = 0;
 	p = (unsigned char *)buffer;
